+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ Unreleased
 * Qblox backend - Fix combination of cluster and latency corrections when converting hw_configs to new specs  (!417)
 * Qblox backend - Fix handling of composite pulses (#299, !411)
 * Qblox backend - Implementation of distortion correction (#285, !388)
+* Qblox backend - Fix incompatibility of distortion_correction parameters as numpy arrays (!426)
 * Operations - Added a `symmetric` key in the `gate_info` to flag symmetric operations. (!389)
 * Operations - Introduce basic CZ-gate via `CompositeSquareEdge` (utilizing `quantify_scheduler.operations.pulse_factories.composite_square_pulse`); replaces the incomplete `SuddenNetZeroEdge` basic CZ-gate implementation (!411)
 * Helpers - Moved MockLocalOscillator definition from tests to `helpers.mock_instruments.MockLocalOscillator` (!392, !336).